jobs: introduce jobspb.JobID
This commit introduces a `jobspb.JobID` int64 type and uses it in most
places where we were previously using an int64.

Release justification: Low-risk change to existing functionality.

Release note: None

+++ pkg/sql/table.go
@@ -149,7 +149,7 @@ func (p *planner) createOrUpdateSchemaChangeJob(
 		// TODO (lucy): get rid of this when we get rid of MutationJobs.
 		if mutationID != descpb.InvalidMutationID {
 			tableDesc.MutationJobs = append(tableDesc.MutationJobs, descpb.TableDescriptor_MutationJob{
-				MutationID: mutationID, JobID: newJob.ID()})
+				MutationID: mutationID, JobID: int64(newJob.ID())})
 		}
 		log.Infof(ctx, "queued new schema change job %d for table %d, mutation %d",
 			newJob.ID(), tableDesc.ID, mutationID)
@@ -180,7 +180,7 @@ func (p *planner) createOrUpdateSchemaChangeJob(
 				// Also add a MutationJob on the table descriptor.
 				// TODO (lucy): get rid of this when we get rid of MutationJobs.
 				tableDesc.MutationJobs = append(tableDesc.MutationJobs, descpb.TableDescriptor_MutationJob{
-					MutationID: mutationID, JobID: job.ID()})
+					MutationID: mutationID, JobID: int64(job.ID())})
 			}
 		}
 		if err := job.SetDetails(ctx, p.txn, newDetails); err != nil {
